Workaround for vim terminal popup cursor jump bug
Vim patch 8.2.0413 enables BufFilePre and BufFilePost autocmds for
terminal buffers and also introduces a minor bug affecting nested
autocmds, which causes the cursor to jump to the start of the current
buffer when projectionist is enabled and a terminal popup window is
created, for example by either the FZF.vim or vim-floaterm plugins.

More information:
https://github.com/vim/vim/issues/5820
https://github.com/junegunn/fzf.vim/issues/1164
https://github.com/tpope/vim-projectionist/issues/155

Though this bug has now been fixed in vim, it may take a while to make
its way though to package repositories and onto peoples' machines, so
this commit works around the issue for now by simply avoiding calling
ProjectionistDetect for terminal buffers. It does this by checking that
the autocmd file name refers to a file that actually exists rather than
checking that the &buftype is "terminal" because the &buftype has not
been set to "terminal" at the time the BufFilePost autocmd is triggered.